reorganize files structure
