+++ catalog/app/containers/Bucket/Dir.tsx
@@ -18,7 +18,7 @@ import * as AssistantContext from './DirAssistantContext'
 import DndWrapper from './DndWrapper'
 import * as Listing from './Listing'
 import * as PD from './PackageDialog'
-import * as FI from './PackageDialog/FilesInput'
+import * as FI from './PackageDialog/Inputs/Files/State'
 import * as Selection from './Selection'
 import Summary from './Summary'
 import * as DirToolbar from './Dir/Toolbar'
