+++ catalog/app/containers/Bucket/Dir/Toolbar/Add/UploadDialog.tsx
@@ -5,8 +5,8 @@ import * as M from '@material-ui/core'
 import * as Lab from '@material-ui/lab'
 
 import { useUploads } from 'containers/Bucket/PackageDialog/Uploads'
-import * as FI from 'containers/Bucket/PackageDialog/FilesState'
-import { FilesInput } from 'containers/Bucket/PackageDialog/FilesInput'
+import * as FI from 'containers/Bucket/PackageDialog/Inputs/Files/State'
+import { FilesInput } from 'containers/Bucket/PackageDialog/Inputs/Files/Input'
 import type * as Toolbar from 'containers/Bucket/Toolbar'
 import Log from 'utils/Logging'
 import * as s3paths from 'utils/s3paths'
