+++ catalog/app/containers/Bucket/PackageDialog/Layout.spec.ts
@@ -1,25 +1,25 @@
-import * as PD from './PackageDialog'
+import { calcDialogHeight } from './Layout'
 
 jest.mock('constants/config', () => ({}))
 
-describe('containers/Bucket/PackageDialog/PackageDialog', () => {
+describe('containers/Bucket/PackageDialog/Layout', () => {
   describe('calcDialogHeight', () => {
     it('height should be minimal for small screen', () => {
       const metaHeight = 0
       const windowHeight = 100
-      expect(PD.calcDialogHeight(windowHeight, metaHeight)).toBe(420)
+      expect(calcDialogHeight(windowHeight, metaHeight)).toBe(420)
     })
 
     it('height should fit into normal screen', () => {
       const metaHeight = 300
       const windowHeight = 768
-      expect(PD.calcDialogHeight(windowHeight, metaHeight)).toBe(568)
+      expect(calcDialogHeight(windowHeight, metaHeight)).toBe(568)
     })
 
     it('height should be enough for content on large screen', () => {
       const metaHeight = 300
       const windowHeight = 1440
-      expect(PD.calcDialogHeight(windowHeight, metaHeight)).toBe(645)
+      expect(calcDialogHeight(windowHeight, metaHeight)).toBe(645)
     })
   })
 })
