+++ catalog/app/containers/Bucket/PackageDialog/index.ts
@@ -1,13 +1,10 @@
 export { default as DialogError } from './DialogError'
 export { default as DialogLoading } from './DialogLoading'
 export { default as DialogSuccess } from './DialogSuccess'
-export * from './FilesInput'
 export * as Inputs from './Inputs'
 export * from './Layout'
 export * from './Manifest'
-export * from './MetaInput'
 export * from './PackageCreationForm'
-export * from './PackageDialog'
 export * from './Skeleton'
 export { useContext, Provider, isPackageHandle } from './State'
 export { default as SubmitSpinner } from './SubmitSpinner'
